requirements.txt
streamlit==1.55.0
pandas==2.3.3
openai==2.29.0
python-dotenv==1.2.2
plotly==6.6.0
